Added logic to set default props for Android

+++ android/src/oldarch/java/com/oblador/shimmer/RNShimmerManager.java
@@ -53,7 +53,7 @@ public class RNShimmerManager extends ViewGroupManager<RNShimmeringView> {
     }
 
     @ReactProp(name = "tilt", defaultFloat = 0.0f)
-    public void setTilt(RNShimmeringView view, int value) {
+    public void setTilt(RNShimmeringView view, float value) {
         RNShimmerManagerImpl.setTilt(view, value);
     }
